requirements.txt
streamlit==1.33.0
passlib==1.7.4
mysql-connector==2.2.9
mysql-connector-python==8.4.0
google-generativeai==0.7.0
requests==2.31.0
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2